added new Level
