+++ index.html
@@ -15,6 +15,6 @@
       <p id="lifes">1</p>
       <p id="score">2</p>
     </footer>
-    <script src="app.js"></script>
+    <script type="module" src="app.js"></script>
   </body>
 </html>
